+++ changelog.md
@@ -1,5 +1,26 @@
 # Changelog for [`deepseq` package](http://hackage.haskell.org/package/deepseq)
 
+## 1.4.7.0
+
+  * Add instances for `Solo` (GHC-9)
+    ([#69](https://github.com/haskell/deepseq/pull/69))
+  * Add once again `infixr 0 deepseq`
+    ([#56](https://github.com/haskell/deepseq/pull/56), [#74](https://github.com/haskell/deepseq/issues/74))
+  * Add `NFData` instance for `ByteArray`
+    ([#65](https://github.com/haskell/deepseq/pull/65))
+
+## 1.4.6.1
+
+  * Revert `infixr 0 deepseq`; this does not appear in the version of `deepseq` pinned to GHC 9.2.1
+
+## 1.4.6.0
+
+  * Bundled with GHC 9.2.1
+  * Remove instances for Data.Semigroup.Option for GHC >= 9.2
+    ([#62](https://github.com/haskell/deepseq/pull/62))
+  * Set the `infixr 0 deepseq` to be consistent with `seq`
+    ([#56](https://github.com/haskell/deepseq/pull/56))
+
 ## 1.4.5.0
 
   * Add `GNFData` for URec